cleanup (commented code, comparison warnings)
Former-commit-id: 795badebc8309439da8b96fb3d34b0d541949e22

+++ src/model/ISignal.cpp
@@ -228,9 +228,9 @@ MsgHandler::msgStatus ISignal::resetMsg (const IMessage* msg)
 //--------------------------------------------------------------------------
 MsgHandler::msgStatus ISignal::defaultMsg (const IMessage* msg)
 {
-	if (msg->size() ==  (unsigned int)(ParallelSignal::dimension()) ) {
+	if (msg->size() ==  ParallelSignal::dimension() ) {
 		std::vector<float> defaultValue;
-		for ( unsigned int i = 0 ; i < msg->size() ; i++ )
+		for ( int i = 0 ; i < msg->size() ; i++ )
 		{
 			float val;
 			if (msg->param(i, val)) {
